Use float instead of int for points

+++ exercise.go
@@ -4,7 +4,7 @@ type ExerciseService service
 
 type Correction struct {
 	Student		string	`yaml:"student"`
-	Points		int		`yaml:"points"`
+	Points		float64	`yaml:"points"`
 	Corrected	bool	`yaml:"corrected"`
 	Correction	string	`yaml:"correction,flow"`
 }
